fix pom
To be able to make a fat jar that includes dependencies it was necessary to remove the system scope from the OpenMPI dependency. 

To do this the jar file had to be installed to a local repository.

In order to make this procedure portable this installation must happen automatically by including mpi.jar from the new "lib" directory.

Anyone who wished to build a package of this project will have to place the mpi.jar from the HPC cluster or computer to be used.
fix assembly


removed useless tag