fix scatterEqually
It will now work as expected when there elements of the sent array are not equally divisible by the number of nodes.

The implementation currently uses MPI_Send and MPI_Recv. ISend and IRecv could be used instead.

A subset of the array to send is copied, if there is an other way that would avoid copying it should be preferred to improve performance.